feat(security): Implement dev auto-login via event subscriber
Replaces the previous authenticator-based approach with a more robust event subscriber to handle automatic login in the development environment.

The previous implementation caused `ClassNotFoundException` errors in the user's environment. This new approach listens for the `kernel.request` event and programmatically logs in the default admin user (ID 1) before the security firewall can deny access. This method is more direct and less prone to issues with the authenticator manager.
