+++ config/packages/security.yaml
@@ -53,10 +53,3 @@ when@test:
                 cost: 4 # Lowest possible value for bcrypt
                 time_cost: 3 # Lowest possible value for argon
                 memory_cost: 10 # Lowest possible value for argon
-
-when@dev:
-    security:
-        firewalls:
-            main:
-                custom_authenticators:
-                    - App\Security\DevAutoLoginAuthenticator
